Trait beefy_gadget_rpc::BeefyApi

pub trait BeefyApi<Notification, Hash>: Sized + Send + Sync + 'static {
    type Metadata: PubSubMetadata;
    fn subscribe_justifications(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<Notification>
    );

    fn unsubscribe_justifications(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool>;

    fn latest_finalized(&self) -> BoxFuture<Result<Hash, Error>>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        Notification: Send + Sync + 'static + Serialize,
        Hash: Send + Sync + 'static + Serialize,
    { ... }
}

Provides RPC methods for interacting with BEEFY.

Associated Types

type Metadata: PubSubMetadata

RPC Metadata

Required methods

fn subscribe_justifications(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<Notification>
)

Returns the block most recently finalized by BEEFY, alongside side its justification.

fn unsubscribe_justifications(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool>

Unsubscribe from receiving notifications about recently finalized blocks.

fn latest_finalized(&self) -> BoxFuture<Result<Hash, Error>>

Returns hash of the latest BEEFY finalized block as seen by this client.

The latest BEEFY block might not be available if the BEEFY gadget is not running in the network or if the client is still initializing or syncing with the network. In such case an error would be returned.

Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
Notification: Send + Sync + 'static + Serialize,
Hash: Send + Sync + 'static + Serialize,

Create an IoDelegate, wiring rpc calls to the trait methods.