Trait beefy_gadget_rpc::BeefyApiClient

pub trait BeefyApiClient<Notification, Hash>: SubscriptionClientT where
    Notification: Send + Sync + 'static + DeserializeOwned,
    Hash: Send + Sync + 'static + DeserializeOwned, {
    fn latest_finalized<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<Hash>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn subscribe_justifications<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Subscription<Notification>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Client implementation for the BeefyApi RPC API.

Provided methods

fn latest_finalized<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = RpcResult<Hash>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Returns hash of the latest BEEFY finalized block as seen by this client.

The latest BEEFY block might not be available if the BEEFY gadget is not running in the network or if the client is still initializing or syncing with the network. In such case an error would be returned.

fn subscribe_justifications<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Subscription<Notification>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Returns the block most recently finalized by BEEFY, alongside side its justification.